Conflicts:
	Sources/Internal/Autotesting/AutotestingDB.cpp
	Sources/Internal/Autotesting/AutotestingDB.h
	Sources/Internal/Autotesting/AutotestingSystem.cpp
	Sources/Internal/Autotesting/AutotestingSystemLua.cpp
	Sources/Internal/Base/RefPtr.h
	Sources/Internal/Core/Core.cpp
	Sources/Internal/Core/Core.h
	Sources/Internal/Entity/Component.cpp
	Sources/Internal/FileSystem/VariantType.cpp
	Sources/Internal/Particles/ParticleLayer.cpp
	Sources/Internal/Particles/ParticleLayer.h
	Sources/Internal/Scene3D/Components/LodComponent.h
	Sources/Internal/Scene3D/Components/ParticleEffectComponent.cpp
	Sources/Internal/UI/UISlider.cpp
	Sources/Internal/UI/UITextFieldAndroid.h
	Sources/Internal/Utils/Random.h
	Tools/ResourceEditor/Classes/CommandLine/Dump/SceneDumper.cpp
	Tools/ResourceEditor/Classes/CommandLine/TextureDescriptor/TextureDescriptorUtils.cpp
	Tools/ResourceEditor/Classes/Commands2/CustomColorsCommands2.cpp
	Tools/ResourceEditor/Classes/Commands2/HeightmapEditorCommands2.cpp
	Tools/ResourceEditor/Classes/Commands2/ParticleEditorCommands.h
	Tools/ResourceEditor/Classes/Commands2/SaveEntityAsAction.cpp
	Tools/ResourceEditor/Classes/Qt/DockLandscapeEditorControls/LandscapeEditorPanels/TilemaskEditorPanel.cpp
	Tools/ResourceEditor/Classes/Qt/DockParticleEditor/EmitterLayerWidget.cpp
	Tools/ResourceEditor/Classes/Qt/DockParticleEditor/ParticleEffectPropertiesWidget.cpp
	Tools/ResourceEditor/Classes/Qt/DockParticleEditor/ParticleEmitterPropertiesWidget.cpp
	Tools/ResourceEditor/Classes/Qt/DockProperties/PropertyEditor.cpp
	Tools/ResourceEditor/Classes/Qt/DockProperties/PropertyEditor.h
	Tools/ResourceEditor/Classes/Qt/DockSceneInfo/SceneInfo.cpp
	Tools/ResourceEditor/Classes/Qt/DockSceneTree/SceneTree.cpp
	Tools/ResourceEditor/Classes/Qt/DockSceneTree/SceneTree.h
	Tools/ResourceEditor/Classes/Qt/DockSceneTree/SceneTreeItem.cpp
	Tools/ResourceEditor/Classes/Qt/DockSceneTree/SceneTreeModel.cpp
	Tools/ResourceEditor/Classes/Qt/Main/ModificationWidget.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/SceneEditor2.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/SceneEditor2.h
	Tools/ResourceEditor/Classes/Qt/Scene/SceneHelper.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/CollisionSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/CollisionSystem.h
	Tools/ResourceEditor/Classes/Qt/Scene/System/CollisionSystem/CollisionLandscape.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/CustomColorsSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/CustomColorsSystem.h
	Tools/ResourceEditor/Classes/Qt/Scene/System/EditorParticlesSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/HeightmapEditorSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/HeightmapEditorSystem.h
	Tools/ResourceEditor/Classes/Qt/Scene/System/HoodSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/HoodSystem/MoveHood.h
	Tools/ResourceEditor/Classes/Qt/Scene/System/HoodSystem/RotateHood.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/HoodSystem/ScaleHood.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorDrawSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorDrawSystem/LandscapeProxy.h
	Tools/ResourceEditor/Classes/Qt/Scene/System/ModifSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/RulerToolSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/SelectionSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/StructureSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/TilemaskEditorSystem.cpp
	Tools/ResourceEditor/Classes/Qt/TextureBrowser/TextureBrowser.cpp
	Tools/ResourceEditor/Classes/Qt/TextureBrowser/TextureConvertor.cpp
	Tools/ResourceEditor/Classes/Qt/TextureBrowser/TextureListDelegate.cpp
	Tools/ResourceEditor/Classes/Qt/TextureBrowser/TextureProperties.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/AddSwitchEntityDialog/AddSwitchEntityDialog.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataDavaKeyedArchive.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataDavaKeyedArchive.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataDavaVariant.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataDavaVariant.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataInspColl.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataInspColl.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataInspDynamic.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataInspMember.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataIntrospection.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataIntrospection.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataKeyedArchiveMember.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyData/QtPropertyDataMetaObject.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyEditor.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyModel.h
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyModelFiltering.h
	Tools/ResourceEditor/Classes/Qt/Tools/TileTexturePreviewWidget/TileTexturePreviewWidget.cpp
	Tools/ResourceEditor/Classes/StringConstants.h

Former-commit-id: 18e757bfa04daac72b92414f82cff62edc4e0d65

+++ Sources/Internal/MemoryManager/MemoryManager.cpp
@@ -35,6 +35,7 @@
 
 #if defined(__DAVAENGINE_WIN32__)
 #include <dbghelp.h>
+#elif defined(__DAVAENGINE_WIN_UAP__)
 #elif defined(__DAVAENGINE_MACOS__) || defined(__DAVAENGINE_IPHONE__)
 #include <execinfo.h>
 #include <dlfcn.h>
@@ -219,7 +220,7 @@ DAVA_NOINLINE void* MemoryManager::Allocate(size_t size, uint32 poolIndex)
         block->allocTotal = static_cast<uint32>(MallocHook::MallocSize(block->realBlockStart));
         block->bktraceHash = 0;
         if (0 == block->allocTotal)
-            block->allocTotal = totalSize;
+            block->allocTotal = static_cast<uint32>(totalSize);
 
         if (tlsAllocScopeStack.IsCreated())
         {
@@ -288,7 +289,7 @@ DAVA_NOINLINE void* MemoryManager::AlignedAllocate(size_t size, size_t align, ui
         block->allocTotal = static_cast<uint32>(MallocHook::MallocSize(block->realBlockStart));
         block->bktraceHash = 0;
         if (0 == block->allocTotal)
-            block->allocTotal = totalSize;
+            block->allocTotal = static_cast<uint32>(totalSize);
 
         if (tlsAllocScopeStack.IsCreated())
         {
@@ -402,7 +403,7 @@ void* MemoryManager::InternalAllocate(size_t size)
         block->allocByApp = static_cast<uint32>(size);
         block->allocTotal = static_cast<uint32>(MallocHook::MallocSize(block));
         if (0 == block->allocTotal)
-            block->allocTotal = totalSize;
+            block->allocTotal = static_cast<uint32>(totalSize);
 
         // Update stat
         {
